+++ app/Observers/AttributeValueObserver.php
@@ -10,12 +10,12 @@ class AttributeValueObserver
     public function creating(AttributeValue $attributeValue): void
     {
         $attributeValue->slug = str($attributeValue->name)->slug()->toString();
-        $attributeValue->created_by = Auth::id();
+        $attributeValue->created_by = (int) Auth::id();
     }
 
     public function updating(AttributeValue $attributeValue): void
     {
         $attributeValue->slug = str($attributeValue->name)->slug()->toString();
-        $attributeValue->updated_by = Auth::id();
+        $attributeValue->updated_by = (int) Auth::id();
     }
 }
