+++ app/Observers/ExpenseCategoryObserver.php
@@ -9,12 +9,12 @@ class ExpenseCategoryObserver
     public function creating(ExpenseCategory $expenseCategory): void
     {
         $expenseCategory->slug = str($expenseCategory->name)->slug()->toString();
-        $expenseCategory->created_by = auth()->id();
+        $expenseCategory->created_by = (int) auth()->id();
     }
 
     public function updating(ExpenseCategory $expenseCategory): void
     {
         $expenseCategory->slug = str($expenseCategory->name)->slug()->toString();
-        $expenseCategory->updated_by = auth()->id();
+        $expenseCategory->updated_by = (int) auth()->id();
     }
 }
