+++ app/Observers/ExpenseObserver.php
@@ -8,11 +8,11 @@ class ExpenseObserver
 {
     public function creating(Expense $expense): void
     {
-        $expense->created_by = auth()->id();
+        $expense->created_by = (int) auth()->id();
     }
 
     public function updating(Expense $expense): void
     {
-        $expense->updated_by = auth()->id();
+        $expense->updated_by = (int) auth()->id();
     }
 }
