+++ app/Observers/UnitTypeObserver.php
@@ -8,13 +8,13 @@ class UnitTypeObserver
 {
     public function creating(UnitType $unitType): void
     {
-        $unitType->created_by = auth()->id();
+        $unitType->created_by = (int) auth()->id();
         $unitType->slug = str($unitType->name)->slug()->toString();
     }
 
     public function updating(UnitType $unitType): void
     {
         $unitType->slug = str($unitType->name)->slug()->toString();
-        $unitType->updated_by = auth()->id();
+        $unitType->updated_by = (int) auth()->id();
     }
 }
